auto_news_intelligence/requirements.txt
sentence-transformers
scikit-learn
numpy
openai
streamlit
plotly
beautifulsoup4
lxml
python-dotenv
pandas
